Replace removed with_min_level method in android_logger
The android_logger crate marked the with_min_level method in version
in 0.12.0 and removed it in 0.13.0. Because futuresdr is currently
using "0.13" of the crate, this breaks the Android build with
"no method named with_min_level` found". Also see
https://github.com/rust-mobile/android_logger-rs/pull/62

This replaces the method call with the new "with_max_level".